+++ skyfield/tests/test_planetarylib.py
@@ -228,6 +228,28 @@ def test_observing_earth_from_location_on_moon_with_time_vector():
     arcsecond = 1.0 / 3600.0
     assert abs(np.array(want) - pair).max() < 12 * arcsecond
 
+def test_using_elevation_to_locate_center_of_moon():
+    # Test the `elevation_m` parameter by using it to offset a Moon
+    # surface location back to the Moon's center.
+
+    ts = load.timescale(builtin=True)
+    t = ts.utc(2020, 4, 23)
+
+    eph = load('de421.bsp')
+    a1 = eph['moon'].at(t)
+
+    pc = PlanetaryConstants()
+    pc.read_text(load('moon_080317.tf'))
+    pc.read_text(load('pck00008.tpc'))
+    pc.read_binary(load('moon_pa_de421_1900-2050.bpc'))
+    frame = pc.build_frame_named('MOON_ME_DE421')
+
+    place = pc.build_latlon_degrees(frame, 26.3, 313.2, -1737400)
+    a2 = (eph['moon'] + place).at(t)
+
+    mm = 1e-3
+    assert max(abs(a1.position.m - a2.position.m)) < mm
+
 def test_frame_alias():
     pc = PlanetaryConstants()
     pc.read_text(load('moon_080317.tf'))